+++ R/proj.R
@@ -1,5 +1,13 @@
 proj <- new.env(parent = emptyenv())
 
+proj_get_ <- function() proj$cur
+
+proj_set_ <- function(path) {
+  old <- proj$cur
+  proj$cur <- path
+  invisible(old)
+}
+
 #' Utility functions for the active project
 #'
 #' Most `use_*()` functions act on the **active project**. If it is
@@ -10,10 +18,10 @@ proj <- new.env(parent = emptyenv())
 #' file. It then stores the active project for use for the remainder of the
 #' session.
 #'
-#' In general, end user code should not call `usethis::proj_get()`,
-#' `usethis::proj_set()`, or `usethis::proj_path()`. They are internal functions
-#' that are exported for occasional interactive use or use in packages that
-#' extend usethis. End user code should call functions in
+#' In general, end user scripts should not contain direct calls to
+#' `usethis::proj_*()` utility functions. They are internal functions that are
+#' exported for occasional interactive use or use in packages that extend
+#' usethis. End user code should call functions in
 #' [rprojroot](https://rprojroot.r-lib.org) or its simpler companion,
 #' [here](https://here.r-lib.org), to programmatically detect a project and
 #' build paths within it.
@@ -31,19 +39,25 @@ proj <- new.env(parent = emptyenv())
 #' ## build a path within the active project (both produce same result)
 #' proj_path("R/foo.R")
 #' proj_path("R", "foo", ext = "R")
+#'
+#' ## build a path within SOME OTHER project
+#' with_project("path/to/some/other/project", proj_path("blah.R"))
+#'
+#' ## convince yourself that with_project() temporarily changes the project
+#' with_project("path/to/some/other/project", print(proj_sitrep()))
 #' }
 NULL
 
-#' @describeIn proj_utils Retrieves the active project and, if necessary, attempts to set it in the first place.
-#' @param quiet Logical. Whether to announce project activation.
+#' @describeIn proj_utils Retrieves the active project and, if necessary,
+#'   attempts to set it in the first place.
 #' @export
-proj_get <- function(quiet = FALSE) {
+proj_get <- function() {
   # Called for first time so try working directory
   if (!proj_active()) {
-    proj_set(".", quiet = quiet)
+    proj_set(".")
   }
 
-  proj$cur
+  proj_get_()
 }
 
 #' @describeIn proj_utils Sets the active project.
@@ -54,40 +68,71 @@ proj_get <- function(quiet = FALSE) {
 #'   project-signalling infrastructure, such as initialising a Git repo or
 #'   adding a DESCRIPTION file.
 #' @export
-proj_set <- function(path = ".", force = FALSE, quiet = FALSE) {
-  if (!is.null(path)) {
-    check_is_dir(path)
-  }
-  path <- proj_path_prep(path)
-
-  if (!force) {
-    new_project <- proj_path_prep(proj_find(path))
-    if (is.null(new_project)) {
-      stop_glue(
-        "Path {value(path)} does not appear to be inside a project or package."
-      )
-    }
-    path <- new_project
+proj_set <- function(path = ".", force = FALSE) {
+  if (is.null(path) || force) {
+    path <- proj_path_prep(path)
+    proj_string <- if (is.null(path)) code("NULL") else value(path)
+    done("Setting active project to {proj_string}")
+    return(proj_set_(path))
   }
 
-  proj_set_(path, quiet = quiet)
+  check_is_dir(path)
+  new_project <- proj_find(path)
+  if (is.null(new_project)) {
+    stop_glue(
+      "Path {value(path)} does not appear to be inside a project or package."
+    )
+  }
+  proj_set(path = new_project, force = TRUE)
 }
 
-#' @describeIn proj_utils Builds a path within the active project. Thin wrapper
-#'   around [fs::path()].
+#' @describeIn proj_utils Builds a path within the active project returned by
+#'   `proj_get()`. Thin wrapper around [fs::path()].
 #' @inheritParams fs::path
 #' @export
 proj_path <- function(..., ext = "") {
   path_norm(path(proj_get(), ..., ext = ext))
 }
 
-proj_set_ <- function(path, quiet = FALSE) {
-  old <- proj$cur
-  proj$cur <- path
-  if (!quiet) {
-    done("Setting active project to {value(proj$cur)}")
-  }
-  invisible(old)
+#' @describeIn proj_utils Runs code with a temporary active project. It is an
+#'   example of the `with_*()` functions in [withr](http://withr.r-lib.org).
+#' @param code Code to run with temporary active project.
+#' @param quiet Whether to suppress user-facing messages, while operating in the
+#'   temporary active project.
+#' @export
+with_project <- function(path = ".",
+                         code,
+                         force = FALSE,
+                         quiet = getOption("usethis.quiet", default = FALSE)) {
+  old_quiet <- options(usethis.quiet = quiet)
+  old_proj  <- proj_set(path = path, force = force)
+
+  on.exit({
+    proj_set(path = old_proj, force = TRUE)
+    options(old_quiet)
+  })
+
+  force(code)
+}
+
+#' @describeIn proj_utils Sets an active project until the current execution
+#'   environment goes out of scope, e.g. the end of the current function or
+#'   test.  It is an example of the `local_*()` functions in
+#'   [withr](http://withr.r-lib.org).
+#' @param .local_envir The environment to use for scoping. Defaults to current
+#'   execution environment.
+#' @export
+local_project <- function(path = ".",
+                          force = FALSE,
+                          quiet = getOption("usethis.quiet", default = FALSE),
+                          .local_envir = parent.frame()) {
+  old_quiet <- options(usethis.quiet = quiet)
+  old_proj  <- proj_set(path = path, force = force)
+
+  withr::defer({
+    proj_set(path = old_proj, force = TRUE)
+    options(old_quiet)
+  }, envir = .local_envir)
 }
 
 ## usethis policy re: preparation of the path to active project
@@ -152,7 +197,7 @@ check_is_package <- function(whos_asking = NULL) {
   stop_glue(message)
 }
 
-proj_active <- function() !is.null(proj$cur)
+proj_active <- function() !is.null(proj_get_())
 
 is_in_proj <- function(path) {
   if (!proj_active()) {